requirements.txt
# Project-wide runtime requirements
Flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.30.0
reportlab>=4.0.0
pillow>=10.0.0
pytesseract>=0.3.10
